Fixed logs filtering order; CronLogRemoveService moved to it's own package 'service/remove'; Done some refactoring job
