+++ src/mokka/src/main/java/pl/hycom/mokka/emulator/logs/LogsController.java
@@ -1,25 +1,24 @@
 package pl.hycom.mokka.emulator.logs;
 
-import java.util.List;
-import java.util.Set;
-
-import javax.servlet.http.HttpServletRequest;
-
+import com.fasterxml.jackson.annotation.JsonView;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
-
-import com.fasterxml.jackson.annotation.JsonView;
-
 import pl.hycom.mokka.emulator.logs.model.Log;
 import pl.hycom.mokka.web.json.View;
+
+import javax.servlet.http.HttpServletRequest;
+import java.util.List;
+import java.util.Set;
 
 /**
  * @author Hubert Pruszyński <hubert.pruszynski@hycom.pl>, HYCOM S.A.
  */
+@Slf4j
 @RestController
 @RequestMapping(headers = "x-requested-with=XMLHttpRequest")
 public class LogsController {
