+++ src/mokka/src/main/resources/templates/log/page.html
@@ -16,7 +16,8 @@
 					            <label>Service</label>
 								<md-select ng-model="ctrl.search.path"  ng-click="ctrl.getSetOfLogs()"  ng-enter="ctrl.activateFilter()">
 									<md-option value="">All</md-option>
-									<md-option ng-repeat="path in ctrl.getLogs" value="{{path}}">{{path}}</md-option>
+									<md-option ng-repeat="path in ctrl.setOfLogs" value="{{path}}">{{path}}</md-option>
+									<!--zmienic nazwe tablicy na z getLogs na np. setOfLogs -->
 								</md-select>
 					        </md-input-container>
 
@@ -69,7 +70,7 @@
 
 			<md-card>
 				<md-card-content layout="column">
-						<md-list ng-show="ctrl.logs.length" class="md-dense" flex="">
+						<md-list ng-show="ctrl.logs.length" class="md-dense" flex="" ng-init="ctrl.fetchLogs()" lazy-scroll="ctrl.fetchLogs()">
 							<md-subheader class="md-no-sticky">
 								<md-button class="btn btn-primary" ng-click="ctrl.changeFetchMethod()">
 									<md-tooltip md-direction="top">Scroll down to see more logs</md-tooltip>
